refactor: moved function to pick tasks based on project to utils
